+++ packages/yike-design-ui/index.less
@@ -46,4 +46,4 @@
 @import 'components/tree-select/style/index.less';
 @import 'components/typography/style/index.less';
 @import 'components/upload/style/index.less';
-@import 'components/yk-time-picker/style/index.less';
+@import 'components/time-picker/style/index.less';
